Orden de transparencias
- Las entidades translúcidas se ordenan cada frame
en función de su distancia a cámara. Se pintan de más
lejanas a más cercanas, siguiendo el algoritmo del pintor
- Arreglado pequeño fallo en 'ParticleSystem'

+++ src/Kepri3D/ParticleSystem.h
@@ -56,7 +56,7 @@ private:
 	glm::vec3* m_velocities;
 
 	/* Tiempo que lleva viva cada partícula */
-	GLuint* m_life;
+	float* m_life;
 };
 
 // - - - - - - - - - -
